continued refactor
