+++ cv_next/app/actions/comments/index.ts
@@ -1,0 +1,5 @@
+import { addLikeToComment, removeLikeFromComment } from "./commentLikes";
+import { fetchComments, fetchChildComments } from "./fetchComments";
+import { postComment } from "./postComment";
+
+export { addLikeToComment, removeLikeFromComment, fetchComments, fetchChildComments, postComment };
