+++ cv_next/app/actions/cvs/index.ts
@@ -1,0 +1,4 @@
+import { fetchCv } from "./fetchCv";
+import { fetchCvs, fetchAllCvs } from "./fetchCvs";
+
+export { fetchCv, fetchCvs, fetchAllCvs };
