+++ cv_next/app/actions/users/index.ts
@@ -1,0 +1,3 @@
+import { fetchUser } from "./fetchUser";
+
+export { fetchUser };
